requirements.txt
flask==3.0.0
flask-socketio==5.3.6
pytest==7.4.2
pytube==15.0.0
python-socketio==5.10.0
python-engineio==4.8.0
flask_cors==4.0.0
datetime==5.2
typing-extensions==4.8.0
python-dotenv==1.0.0
pymongo==4.6.0
mongodb-odm==0.2.4